feature: flink job config for jar web
